Updated useRemixForm hook.
Hours of reading package documentation, package alternatives, experimental testing, and I've finally understood the ins and outs of react-hook-form and Remix's useFetcher and clientAction
I coded a beautiful solution and was so lost in it that I realized that the flexibility and control in terms of form submission handling were--in fact-- not needed.
Dang it. At least I learned something.